pin/test-requirements.txt
pytest == 6.1.2
requests==2.18.1
pyyaml==5.3.1
redis==3.5.3
